Adding question 2

+++ 120050033_120050041_120050087_assignment2_description/2/code/myGaussianNoiser.m
@@ -4,5 +4,5 @@ function noisy_image = myGaussianNoiser(image)
     
     standard_deviation = 0.05*(maximum_intensity - minimum_intensity);
     
-    noisy_image = uint8(standard_deviation*randn(size(image, 1), size(image, 2)) + image);
+    noisy_image = standard_deviation*randn(size(image, 1), size(image, 2)) + image;
 end
